Improved prfiles Time Complexity
For the part of the program that handles block printing, the list
of blocks was being sorted in order to compress the output on the
console. Previously, it was using insertion sort because it was
easy to program and was sufficient in early testing. However, it
became quickly apparent that sorting needed to be much faster. This
update switched the function over to using mergesort, which made
the section of the code run literally thousands of times faster.