Added Defragmentation Bonus
I have added the code that I wrote for the bonus. Towards the end of the
assignment, we were given the opportunity to do an extra credit assignment
on a topic of our choosing. I chose to do defragmentation, and wrote a program
that performs it on the filesystem. The goal was to minimize the external
fragmentation of the filesystem, which I was able to do.